Updated Final models
Decided on The 14day SPEI and Tmax additive model. Updated code in the plotting to show only significant t-vals for each variable in Figure A. Still need to track down a data issue in the ndviMet object. We have a weird gap in late February that is causing issues. May be a data input issue.